refactor: move testes de editar perfil usuário para nova classe
