+++ test/UsuarioControllerTest.cs
@@ -15,12 +15,11 @@ using app.Entidades;
 using app.Controllers;
 using api.Usuarios;
 using api.Senhas;
-using api;
-using app.Services;
+using Xunit.Microsoft.DependencyInjection.Abstracts;
 
 namespace test
 {
-    public class UsuarioControllerTest : AuthTest, IDisposable
+    public class UsuarioControllerTest : TestBed<Base>, IDisposable
     {
         const int CREATED = 201;
         const int INTERNAL_SERVER_ERROR = 500;
@@ -34,7 +33,7 @@ namespace test
             dbContext.PopulaUsuarios(5);
         }
 
-        public async Task<(string Token, string TokenAtualizacao)> AutenticarUsuarioLocal(UsuarioDTO usuario)
+        public async Task<(string Token, string TokenAtualizacao)> AutenticarUsuario(UsuarioDTO usuario)
         {
             var resultado = await controller.Logar(usuario);
 
@@ -70,7 +69,7 @@ namespace test
         {
             var usuario = dbContext.PopulaUsuarios(1, includePerfil: true).First();
 
-            await AutenticarUsuarioLocal(usuario);
+            await AutenticarUsuario(usuario);
             var permissoes = await controller.ListarPermissoes();
             Assert.NotEmpty(permissoes);
         }
@@ -80,7 +79,7 @@ namespace test
         {
             var usuario = dbContext.PopulaUsuarios(1, includePerfil: true).First();
 
-            var login = await AutenticarUsuarioLocal(usuario);
+            var login = await AutenticarUsuario(usuario);
             var novoLogin = await controller.AtualizarToken(new AtualizarTokenDto
             {
                 Token = login.Token,
@@ -97,7 +96,7 @@ namespace test
         {
             var usuario = dbContext.PopulaUsuarios(1, includePerfil: true).First();
 
-            var login = await AutenticarUsuarioLocal(usuario);
+            var login = await AutenticarUsuario(usuario);
             var atualizarTokenDto = new AtualizarTokenDto
             {
                 Token = login.Token,
@@ -282,160 +281,6 @@ namespace test
 
             usuarioServiceMock.Verify(service => service.TrocaSenha(redefinicaoSenhaDTO), Times.Once);
             Assert.IsType<NotFoundResult>(resultado);
-        }
-
-        [Fact]
-        public async void ObterUsuariosAsync_QuandoNãoTemPermissaoVisualizar_RetornaErroDePermissao()
-        {
-            var ex = await Assert.ThrowsAsync<AuthForbiddenException>(async () =>
-                await controller.ListarAsync(new PesquisaUsuarioFiltro()));
-
-            Assert.Contains("não tem a permissão: Visualizar Usuário", ex.Message);
-        }
-
-        [Fact]
-        public async void ObterUsuariosAsync_QuandoFiltroVazio_RetornaTodosUsuarios()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioVisualizar });
-            var filtro = new PesquisaUsuarioFiltro
-            {
-                ItemsPorPagina = 10,
-            };
-            var usuarios = await controller.ListarAsync(filtro);
-            Assert.Equal(5, usuarios.Total);
-        }
-
-        [Fact]
-        public async void ObterUsuariosAsync_QuandoFiltradoPorUf_RetornaUsuariosDaUfDada()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioVisualizar });
-            var filtro = new PesquisaUsuarioFiltro
-            {
-                UfLotacao = UF.DF,
-            };
-            var u = dbContext.Usuario.ToList();
-            u[0].UfLotacao = UF.DF;
-            u[1].UfLotacao = UF.DF;
-            u[2].UfLotacao = UF.DF;
-            u[3].UfLotacao = UF.AM;
-            u[4].UfLotacao = UF.AM;
-            dbContext.SaveChanges();
-
-            var lista = await controller.ListarAsync(filtro);
-
-            Assert.Equal(UF.DF, lista.Items[0].UfLotacao);
-            Assert.Equal(UF.DF, lista.Items[1].UfLotacao);
-            Assert.Equal(UF.DF, lista.Items[2].UfLotacao);
-            Assert.Equal(3, lista.Items.Count);
-        }
-
-        [Fact]
-        public async void ObterUsuariosAsync_QuandoFiltradoPorMunicipio_RetornaUsuariosDoMunicipioDado()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioVisualizar });
-            var m = new Municipio { Id = 1, Nome = "Municipio", Uf = UF.DF };
-            dbContext.Municipio.Add(m);
-            var filtro = new PesquisaUsuarioFiltro
-            {
-                MunicipioId = m.Id,
-            };
-            var u = dbContext.Usuario.ToList();
-            u[0].MunicipioId = 1;
-            u[1].MunicipioId = 1;
-            u[2].MunicipioId = 2;
-            u[3].MunicipioId = 2;
-            u[4].MunicipioId = 2;
-            dbContext.SaveChanges();
-
-            var lista = await controller.ListarAsync(filtro);
-
-            Assert.Equal(1, lista.Items[0].Municipio!.Id);
-            Assert.Equal(1, lista.Items[1].Municipio!.Id);
-            Assert.Equal(2, lista.Items.Count);
-        }
-
-        [Fact]
-        public async void ObterUsuariosAsync_QuandoFiltradoPorPerfil_RetornaUsuariosComPerfilDado()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioVisualizar });
-            var filtro = new PesquisaUsuarioFiltro
-            {
-                PerfilId = Guid.NewGuid(),
-            };
-            var u = dbContext.Usuario.ToList();
-            u[0].PerfilId = filtro.PerfilId;
-            u[1].PerfilId = filtro.PerfilId;
-            u[2].PerfilId = filtro.PerfilId;
-            dbContext.SaveChanges();
-
-            var lista = await controller.ListarAsync(filtro);
-
-            Assert.Equal(3, lista.Items.Count);
-        }
-
-        [Fact]
-        public async Task EditarPerfilUsuario_QuandoNaoTemPermissao_ErroDePermissao()
-        {
-            AutenticarUsuario(controller, permissoes: new());
-            var dto = new EditarPerfilUsuarioDTO { NovoPerfilId = "id" };
-            var ex = await Assert.ThrowsAsync<AuthForbiddenException>(async ()
-                => await controller.EditarPerfilUsuario(1, dto));
-
-            Assert.Contains("não tem a permissão: Editar Perfil", ex.Message);
-        }
-
-        [Fact]
-        public async Task EditarPerfilUsuario_QuandoUsuarioNaoExiste_RetornaNaoEncontrado()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioPerfilEditar });
-            var dto = new EditarPerfilUsuarioDTO { NovoPerfilId = "id" };
-            var ex = await Assert.ThrowsAsync<ApiException>(async ()
-                => await controller.EditarPerfilUsuario(-1, dto));
-
-            Assert.Equal(ErrorCodes.UsuarioNaoEncontrado, ex.Error.Code);
-        }
-
-        [Fact]
-        public async Task EditarPerfilUsuario_QuandoPerfilNaoExiste_RetornaPerfilNaoEncontrado()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioPerfilEditar });
-            var usuarioId = dbContext.Usuario.First().Id;
-            var dto = new EditarPerfilUsuarioDTO { NovoPerfilId = Guid.NewGuid().ToString() };
-            var excecao = await Assert.ThrowsAsync<ApiException>(async ()
-                => await controller.EditarPerfilUsuario(usuarioId, dto));
-
-            Assert.Equal(ErrorCodes.PermissaoNaoEncontrada, excecao.Error.Code);
-        }
-
-        [Fact]
-        public async Task EditarPerfilUsuario_QuandoTemPermissao_DeveAlterarPerfilDoUsuario()
-        {
-            AutenticarUsuario(controller, permissoes: new() { Permissao.UsuarioPerfilEditar });
-            var novoPerfilParaUsuario = new Perfil
-            {
-                Id = Guid.NewGuid(),
-                // Se remover esses parâmetros, os testes acusam que já foi inserido uma row duplicada
-                Nome = "Teste",
-                Tipo = TipoPerfil.Administrador
-            };
-            dbContext.Perfis.Add(novoPerfilParaUsuario);
-            dbContext.SaveChanges();
-            var usuario = dbContext.Usuario.First();
-            var dto = new EditarPerfilUsuarioDTO { NovoPerfilId = novoPerfilParaUsuario.Id.ToString() };
-
-            await controller.EditarPerfilUsuario(usuario.Id, dto);
-
-            var usuarioEditado = dbContext.Usuario.Find(usuario.Id)!;
-
-            Assert.Equal(novoPerfilParaUsuario.Id, usuarioEditado.PerfilId);
-        }
-
-        public new void Dispose()
-        {
-            dbContext.RemoveRange(dbContext.PerfilPermissoes);
-            dbContext.RemoveRange(dbContext.Perfis);
-            dbContext.RemoveRange(dbContext.Usuario);
-            dbContext.RemoveRange(dbContext.Empresa);
         }
     }
 }
